ParlayLib V2 (#13)
ParlayLib V2 adds new features and improvements, including

- The parlay::delayed library, for range operations over delayed sequences
- Several new primitives
- Improved monoid interface
- A new random number generator interface
- Bug fixes, including a potential deadlock in the allocator
- Over 50 example applications for Parlay

+++ include/parlay/hash_table.h
@@ -8,7 +8,6 @@
 #include <iostream>
 
 #include "delayed_sequence.h"
-#include "monoid.h"
 #include "parallel.h"
 #include "sequence.h"
 #include "slice.h"
@@ -64,7 +63,7 @@ class hashtable {
   // Size is the maximum number of values the hash table will hold.
   // Overfilling the table could put it into an infinite loop.
   hashtable(size_t size, HASH hashF, double load = 1.5)
-    : m(100 + static_cast<size_t>(load * size)),
+    : m(100 + static_cast<size_t>(load * static_cast<double>(size))),
       empty(hashF.empty()),
       hashStruct(hashF),
       TA(sequence<eType>(m, empty)) {
